# doc/openapi/modlink-platform.openapi.yaml
openapi: 3.1.0
info:
  title: ModLinkCloud Platform API
  version: "1.4.0"
  description: |
    控制台与管理后台使用的业务 API。与人读文档：`doc/模链云-API接口文档.md`。

    - **命名空间**：统一前缀 **`/mlk`**（共机部署时便于 `location /mlk/` 分流）。
    - **Base**：`https://api.{domain}/mlk/platform/v1`
    - **鉴权**：`Authorization: Bearer <JWT>`；租户上下文可选 **`X-Org-Id`**（优先于 JWT 默认 org）
    - **响应包裹**：`{ code, message, data, request_id }`（`code=0` 成功）
servers:
  - url: https://api.{domain}
    description: 生产（将 {domain} 换为实际备案域名）
security:
  - bearerJwt: []
...
paths:
...
  /mlk/platform/v1/auth/register:
    post:
      tags: [Auth]
      operationId: authRegister
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelopeToken"
  /mlk/platform/v1/auth/login:
    post:
      tags: [Auth]
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelopeToken"
  /mlk/platform/v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: authRefresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelopeToken"
...
  /mlk/platform/v1/auth/password/reset-request:
    post:
      tags: [Auth]
      operationId: authPasswordResetRequest
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                phone:
                  type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
  /mlk/platform/v1/auth/password/reset-confirm:
    post:
      tags: [Auth]
      operationId: authPasswordResetConfirm
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
...
  /mlk/platform/v1/users/me:
    get:
      tags: [Users]
      operationId: usersMeGet
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
    patch:
      tags: [Users]
      operationId: usersMePatch
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                avatar_url:
                  type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
  /mlk/platform/v1/orgs:
    get:
      tags: [Orgs]
      operationId: orgsList
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelopePaged"
    post:
      tags: [Orgs]
      operationId: orgsCreate
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                slug:
                  type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
  /mlk/platform/v1/orgs/{org_id}/switch:
    post:
      tags: [Orgs]
      operationId: orgsSwitch
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
  /mlk/platform/v1/orgs/{org_id}:
    get:
      tags: [Orgs]
      operationId: orgsGet
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
    patch:
      tags: [Orgs]
      operationId: orgsPatch
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
...
  /mlk/platform/v1/api-keys:
    get:
      tags: [ApiKeys]
      operationId: apiKeysList
      parameters:
        - $ref: "#/components/parameters/XOrgId"
        - name: page
          in: query
          schema:
            type: integer
        - name: page_size
          in: query
          schema:
            type: integer
        - name: org_id
          in: query
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelopePaged"
    post:
      tags: [ApiKeys]
      operationId: apiKeysCreate
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, scope]
              properties:
                name:
                  type: string
                scope:
                  type: string
                  enum: [personal, org]
                org_id:
                  type: string
                expires_at:
                  type: string
                  format: date-time
                ip_allowlist:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: data 中含一次性 secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
  /mlk/platform/v1/api-keys/{key_id}:
    patch:
      tags: [ApiKeys]
      operationId: apiKeysPatch
      parameters:
        - $ref: "#/components/parameters/KeyId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
    delete:
      tags: [ApiKeys]
      operationId: apiKeysDelete
      parameters:
        - $ref: "#/components/parameters/KeyId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
...
  /mlk/platform/v1/wallet:
    get:
      tags: [Wallet]
      operationId: walletGet
      parameters:
        - $ref: "#/components/parameters/XOrgId"
        - name: org_id
          in: query
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
...
  /mlk/platform/v1/orders/recharge:
    post:
      tags: [Orders]
      operationId: ordersRechargeCreate
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [amount_cents, channel]
              properties:
                amount_cents:
                  type: integer
                channel:
                  type: string
                  enum: [wechat, alipay]
                org_id:
                  type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
  /mlk/platform/v1/orders:
    get:
      tags: [Orders]
      operationId: ordersList
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: page_size
          in: query
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelopePaged"
  /mlk/platform/v1/orders/{order_id}:
    get:
      tags: [Orders]
      operationId: ordersGet
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
...
  /mlk/platform/v1/usage/summary:
    get:
      tags: [Usage]
      operationId: usageSummary
      parameters:
        - $ref: "#/components/parameters/XOrgId"
        - name: org_id
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: granularity
          in: query
          schema:
            type: string
            enum: [hour, day]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
...
  /mlk/platform/v1/admin/users:
    get:
      tags: [AdminUsers]
      operationId: adminUsersList
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelopePaged"
...
  /mlk/platform/v1/admin/channels:
    get:
      tags: [AdminChannels]
      operationId: adminChannelsList
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelopePaged"
    post:
      tags: [AdminChannels]
      operationId: adminChannelsCreate
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
...
  /mlk/platform/v1/admin/models:
    get:
      tags: [AdminModels]
      operationId: adminModelsList
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelopePaged"
    post:
      tags: [AdminModels]
      operationId: adminModelsCreate
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
...
components:
  securitySchemes:
    bearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XOrgId:
      name: X-Org-Id
      in: header
      required: false
      schema:
        type: string
      description: 指定当前组织（合法成员时优先于 JWT 默认 org）
    OrgId:
      name: org_id
      in: path
      required: true
      schema:
        type: string
...
    KeyId:
      name: key_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ApiEnvelope:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
        data: {}
        request_id:
          type: string
    ApiEnvelopePaged:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items: {}
                page:
                  type: integer
                page_size:
                  type: integer
                total:
                  type: integer
    ApiEnvelopeToken:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                user_id:
                  type: string
                access_token:
                  type: string
                refresh_token:
                  type: string
                expires_in:
                  type: integer
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        invite_code:
          type: string
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        phone:
          type: string
        password:
          type: string